test: split responses local shell coverage
